update phuong thuc tim kiem

+++ ABnB/src/main/java/com/TDD/ABnB/controller/AppBookingController.java
@@ -5,6 +5,7 @@ import com.TDD.ABnB.models.AppBooking;
 import com.TDD.ABnB.models.AppUser;
 import com.TDD.ABnB.services.app_booking_service.AppBookingService;
 import com.TDD.ABnB.services.app_property_service.AppPropertyService;
+import com.TDD.ABnB.services.app_user_service.AppUserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
@@ -15,6 +16,9 @@ import javax.swing.text.DateFormatter;
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
 import java.time.Duration;
+import java.time.LocalDate;
+import java.time.Period;
+import java.time.format.DateTimeFormatter;
 import java.util.Date;
 
 @RestController
@@ -26,6 +30,10 @@ public class AppBookingController {
     private AppBookingService appBookingService;
     @Autowired
     private AppPropertyService appPropertyServiceImpl;
+
+    @Autowired
+    private AppUserService appUserServiceImpl;
+
 
     @GetMapping()
     public ResponseEntity<Iterable<AppBooking>> showListBooking() {
@@ -49,18 +57,16 @@ public class AppBookingController {
         for(AppBooking booking: bookings) {
             Date checkoutDateEx = new SimpleDateFormat("yyyy-MM-dd").parse(booking.getCheckoutDate());
             Date checkinDateEx = new SimpleDateFormat("yyyy-MM-dd").parse(booking.getCheckinDate());
-            boolean bookingDateValid= checkinDate.after(checkoutDate);
-            System.out.println("checkin");
-            System.out.println(checkinDate);
-            System.out.println("checkout");
-            System.out.println(checkoutDateEx);
+            boolean bookingDateValid= checkinDate.before(checkoutDate);
             boolean checkinDateValid= checkinDate.after(checkoutDateEx);
             boolean checkOutDateValid= checkoutDate.before(checkinDateEx);
-            System.out.println(checkinDateValid);
-            if(!bookingDateValid && (!checkinDateValid||!checkOutDateValid)) {
+            if((!bookingDateValid)&&((!checkinDateValid)||(!checkOutDateValid))) {
                 throw new DuplilcateUserException("Booking unavailable");
             }
         }
+        System.out.println(appUserServiceImpl.findById(appBooking.getAppUser().getId()).getAppBookings());
+        System.out.println(appPropertyServiceImpl.findById(appBooking.getAppProperty().getId()).getAppBookings());
+
         appBookingService.save(appBooking);
         ResponseEntity<AppBooking> res=new ResponseEntity<AppBooking>(appBooking, HttpStatus.ACCEPTED);
         return res;
@@ -74,8 +80,18 @@ public class AppBookingController {
     }
 
     @DeleteMapping("/{id}")
-    public void deleteBooking(@PathVariable("id") Long id) {
-        AppBooking appBooking = appBookingService.findById(id);
+    public ResponseEntity<AppUser> deleteBooking(@PathVariable("id") Long id) throws DuplilcateUserException, ParseException {
+        AppBooking appBooking =appBookingService.findById(id);
+        DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd");
+        LocalDate checkinDate= LocalDate.parse(appBooking.getCheckinDate(), formatter);
+        Period period = Period.between(checkinDate, java.time.LocalDate.now());
+        System.out.println(period.getDays());
+        if(period.getDays()<=1) {
+            throw new DuplilcateUserException("Cannot cancel booking this late");
+        }
+        AppUser user= appUserServiceImpl.findById(appBooking.getAppUser().getId());
         appBookingService.delete(appBooking);
+        ResponseEntity<AppUser> res=new ResponseEntity<AppUser>(user,HttpStatus.ACCEPTED);
+        return res;
     }
 }
